refactor: Rename table_utils.py to naming.py for clarity
- Rename core/table_utils.py to core/naming.py for better clarity about purpose
- Update imports in vendors/aws/pipeline.py
- More descriptive name that clearly indicates name sanitization functionality

All tests passing after this change.
